requirements.txt
fastapi==0.115.12
jinja2==3.1.2
python-multipart==0.0.6
mysql-connector-python==8.1.0
pyjwt==2.8.0
python-jose==3.3.0
cryptography==41.0.4
